Update ancientsinglegenome-hostassociated_libraries.tsv
Hope that this time is fine. 
Kocher2021 is not present at all in the raw tsv. file so I have created all the fields from the info found in the paper's supplementary materials. 
However, I could see Kocher2021 in the tsv. file in the "sample" folder. 
AGES: I have compared my field with its field and I see some differences in the calculated ages (100 yrs). I have used the https://nikhausmann.shinyapps.io/BP_to_BC_and_more/ for the conversion and because the ages are indicated as a range for each sample, I have calculated the mean for each range, converted in BP and rounded to century. 
ex. 5277yr-4969yr --> 5123 BC --> 7072 BP --> 7100 BP 
After a checking through these inconsistencies it looks like that the other person who filled the file, if a date was 7249 he/she rounded it to 7300 whether I'd put it as 7200 
DATA TYPE: I put "raw" because they are raw data in the ENA. However, the aligned data are available in the supplementary materials. Let me know what is best to put
ARCHIVE ACC #: I put the "secondary study accession" number