Update-update
Hi James sorry for the mess but I am not familiar with some concepts yet. Thanks for being patient with me.

1) "strand type" column filled with the correct info taken from "library protocol" column in the supplementary material file.
2) I can see the "nominal length" column in the ENA website but I do not where to put this info in our file. Cannot see any column with this name or similar. Sorry!
3) I have update the "download_md5s" column with the correct info taken from the file you sent me. (The download links were ok already). 

Also there are 3 samples that are present in the supplementary material file but not in the ENA website. I have realized it now when filling the "strand type" info. These are: Karsdorf, Petersberg and Sorsum. I guess that since they are not in the ENA database it is fine to leave them out, correct?

Thanks!