Updated Bieker2020
1) Removed the entries previously added for Project PRJNA339123 as these were based on a copy/paste of the full tsv file from SRA and comprised other samples not inputted in our sample tables.
2) Added the 11 samples discussed (10 + USNH-39 which is also associated with another SRA project PRJEB34825)
(FYI here is the information we had discussed: That sample is in our samples table but associated with another archive project PRJEB34825 and its corresponding ERS code ERS3882395. That corresponds to the first sequencing round. The second sequencing round is associated with this archive project (PRJNA339123) and an SRR code SRR4047228. So the sample is present in our table, just not linked to this archive project"