Update ancientsinglegenome-hostassociated_samples.tsv
Long comment warning!!! 

For the sample list I would like to ask, for cross-check on genome type - samples HGC009 and HGC068 clearly state  chromosome in Table 1  (in the publication), and the age of the samples. For the age, I took the reported year span (the same table 1), I took the average year of each date and calculated age from the present. For the missing date (reported: "N/A") of the sample HGC068, I improvised and took the years of Minoan period, calculated the middle point, added that to the average date of all the other samples and divided for the average again. I know this is complex, but this is the averagest of all averages I could think off. Let me know if there's a simpler way. To be fair all three samples with the carbon dating had one closest point of proximity, so should have I use that instead maybe?