Rifkin2023
Lat:Lon was changed to two decimal places, and the data type was changed to raw.
About the tissue type, petrous bone and a premolar were analyzed, so I thought that when referring to multiple tissue types the term should be "tissue". Let me know if this applies to this case.